Update cython code

+++ include/spectral.h
@@ -13,7 +13,7 @@ std::vector<double> temperature_power_law(
 
 std::vector<double> surface_density_profile(
   const std::vector<double> &radius, float inner_radius,
-  float inner_sigma, float p, long long dim);
+  double inner_sigma, float p, long long dim);
 
 std::vector<double> azimuthal_modulation(
   const std::vector<double> &xx, const std::vector<double> &yy,
